backend/requirements.txt
Django==5.1.0
django-ninja==1.3.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-allauth==64.2.0